nilrt_snac: rename OpkgHelper.py to opkg.py
The OpkgHelper.py file has a nonstandard name (capitalized). Rename it
to just lowercase 'opkg' for consistency.
