+++ nilrt_snac/_configs/_ntp_config.py
@@ -5,7 +5,7 @@ from nilrt_snac._configs._base_config import _BaseConfig
 from nilrt_snac._configs._config_file import _ConfigFile
 
 from nilrt_snac import logger
-from nilrt_snac.OpkgHelper import opkg_helper
+from nilrt_snac.opkg import opkg_helper
 
 
 class _NTPConfig(_BaseConfig):
